bugfix: RDF pair distribution for distinct species
Fixes a bug where a RDF pair distribution for distinct species would ignore one of the ij terms in the central periodic cell. This affects the relative peak heights for local coordination shells, and the coordination number calculation.

Includes a test that cubic coordination gives a nearest-neighbour coordination number of 8.